rpc, test: address analyzepsgt review feedback
- Make the dummy-signature size comment say explicitly that 72/73 bytes
  is the *standard low-S* bound. The consensus maximum is one byte more,
  but SCRIPT_VERIFY_LOW_S is in STANDARD_SCRIPT_VERIFY_FLAGS, so a
  high-S signature would not relay and is the wrong bound for a
  broadcast size estimate; the constant is deliberate and unchanged.

- Don't let the negative-fee check in AnalyzePSGT overwrite a per-input
  error recorded earlier (a non-standard input combined with outputs
  exceeding inputs reported only the latter). Add a unit test pinning
  the precedence.

- Push estimated_final_size as int64_t instead of a narrowing (int)
  cast, matching the other numeric fields in the RPC layer.

+++ src/rpc/psgt.cpp
@@ -440,7 +440,7 @@ UniValue analyzepsgt(const UniValue& params)
     result.pushKV("inputs", inputs_arr);
 
     if (analysis.estimated_final_size)
-        result.pushKV("estimated_final_size", (int)*analysis.estimated_final_size);
+        result.pushKV("estimated_final_size", (int64_t)*analysis.estimated_final_size);
     if (analysis.fee)
         result.pushKV("fee", ValueFromAmount(*analysis.fee));
     if (analysis.min_required_fee)
